+++ ICD.Connect.Settings/FileOperations.cs
@@ -9,6 +9,7 @@ using ICD.Connect.Settings.Comparers;
 using ICD.Connect.Settings.Cores;
 using ICD.Connect.Settings.Header;
 using ICD.Connect.Settings.Migration;
+using ICD.Connect.Settings.Utils;
 #if SIMPLSHARP
 using Crestron.SimplSharp.CrestronIO;
 using Activator = Crestron.SimplSharp.Reflection.Activator;