+++ ICD.Connect.Settings/Migration/Migrators/ConfigVersionMigrator_3x1_To_4x0.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using ICD.Common.Utils;
+using ICD.Connect.Settings.Utils;
 #if SIMPLSHARP
 using Crestron.SimplSharp.CrestronXmlLinq;
 #else