+++ ICD.Connect.Settings/PluginFactory.cs
@@ -1,4 +1,5 @@
 ﻿using ICD.Connect.Settings.Attributes;
+using ICD.Connect.Settings.Utils;
 #if SIMPLSHARP
 using Crestron.SimplSharp.Reflection;
 #else